entity link cleanup

+++ modules/flowable-cmmn-engine/src/main/java/org/flowable/cmmn/engine/impl/persistence/entity/CaseInstanceEntityManagerImpl.java
@@ -92,7 +92,9 @@ public class CaseInstanceEntityManagerImpl extends AbstractCmmnEntityManager<Cas
         getIdentityLinkEntityManager().deleteIdentityLinksByScopeIdAndScopeType(caseInstanceId, ScopeTypes.CMMN);
         
         // Entity links
-        getEntityLinkEntityManager().deleteEntityLinksByScopeIdAndScopeType(caseInstanceId, ScopeTypes.CMMN);
+        if (cmmnEngineConfiguration.isEnableEntityLinks()) {
+            getEntityLinkEntityManager().deleteEntityLinksByScopeIdAndScopeType(caseInstanceId, ScopeTypes.CMMN);
+        }
         
         // Tasks
         TaskEntityManager taskEntityManager = getTaskEntityManager();
